pyproject.toml
[project]
name = "robotcourses-api"
version = "0.1.0"
description = "Testes de API com Robot Framework"
authors = [
    {name = "Robot Courses",email = "git@robotcourses.com.br"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "robotframework (>=7.3.2,<8.0.0)",
    "robotframework-requests (>=0.9.7,<0.10.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
